requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.13.1
anyio==4.3.0
apturl==0.5.2
asgiref==3.5.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.1.0
bcrypt==3.2.0
bidict==0.22.1
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
et-xmlfile==1.1.0
exceptiongroup==1.1.2
Faker==19.2.0
fastapi==0.88.0
fasteners==0.14.1
frozenlist==1.3.3
future==0.18.2
fuzzywuzzy==0.18.0
GeoAlchemy2==0.14.1
greenlet==2.0.2
gyp==0.1
h11==0.13.0
httpcore==0.17.3
httplib2==0.20.2
httptools==0.6.0
httpx==0.24.1
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jwt==1.3.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
loguru==0.6.0
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
meld==3.20.4
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.10.6
numpy==1.25.1
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pprintpp==0.4.0
protobuf==3.12.4
psutil==5.9.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pyaml==23.7.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==1.10.11
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-editor==1.0.4
python-engineio==4.5.1
python-magic==0.4.24
python-multipart==0.0.5
python-socketio==5.8.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==4.4.0
reportlab==3.6.8
requests==2.28.1
retrying==1.3.3
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
socketio==0.2.1
SQLAlchemy==2.0.29
ssh-import-id==5.11
starlette==0.22.0
systemd-python==234
tenacity==8.2.3
terminator==2.1.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.20.0
uvloop==0.17.0
vboxapi==1.0
wadllib==1.3.6
watchfiles==0.19.0
websockets==11.0.3
wsproto==1.0.0
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
